% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurv.R
\name{ggsurv}
\alias{ggsurv}
\alias{ggsurv.survfit}
\title{KM Curves with ggplot}
\usage{
ggsurv(s, ...)

\method{ggsurv}{survfit}(s, CI = T, plot.cens = T, surv.col = "gg.def",
  cens.col = "red", lty.est = 1, lty.ci = 2, cens.shape = 3,
  xlab = "Time", ylab = "", main = "", cumProb = F, yTicks = 5,
  dataLabels = "", addCounts = F, bw = F, legend_title, legend_pos, ...)
}
\arguments{
\item{s}{survival object}

\item{CI}{include confidence interval (default TRUE)}

\item{plot.cens}{plot marks to indicate a censored subject (default TRUE)}

\item{surv.col}{colors to use for survival curves. Should match number of groups.}

\item{cens.col}{color of censor marks}

\item{lty.est}{line type for survival curve}

\item{lty.ci}{line type for confidence interval}

\item{cens.shape}{shape of censor marks}

\item{xlab}{label for x axis (default 'Time')}

\item{ylab}{label for y axis (default '')}

\item{main}{title for plot (default '')}

\item{cumProb}{plot cumulative probability instead of survival (default FALSE)}

\item{yTicks}{number of ticks on the y axis}

\item{addCounts}{add number still at risk above the x axis. This is an attempt to replicate SAS behavior. It does not always give good results (default FALSE)}

\item{bw}{print in black and white (default FALSE)}

\item{legend_title}{title to use for the legend}

\item{legend_pos}{where to position the legend}
}
\description{
generic function to create a ggplot kaplan meier curve
}
\examples{
require(survival)
sv1 <- with(ovarian, Surv(futime, fustat))

## fit intercept only survival
sf1 <- survfit(sv1 ~ 1, ovarian)
## basic km curve
ggsurv(sf1)
## cumulative probability instead of survival
ggsurv(sf1, cumProb=T)
## add in survival counts
ggsurv(sf1, cumProb=T, addCounts=T)

## fit model with two strata
sf2 <- survfit(sv1 ~ resid.ds, ovarian)
## basic km curve
ggsurv(sf2, CI=F)
## add in confidence intervals
ggsurv(sf2, CI=T)
## cumulative probability instead of survival
ggsurv(sf2, cumProb=T)
## add in survival counts
ggsurv(sf2, cumProb=T, addCounts=T)
}
\keyword{ggplot}
\keyword{survival}

