% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurv.R
\name{ggsurv.survfit}
\alias{ggsurv.survfit}
\title{ggsurv survfit s3 object}
\usage{
\method{ggsurv}{survfit}(s, CI = T, plot.cens = T, surv.col = "gg.def",
  cens.col = "red", lty.est = 1, lty.ci = 2, cens.shape = 3,
  xlab = "Time", ylab = "", main = "", cumProb = F, yTicks = 5,
  dataLabels = "", addCounts = F, bw = F, legend_title, legend_pos, ...)
}

