% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1_fxn}
\alias{tab1_fxn}
\title{Table Maker


generic function to create a summary table of given variables.}
\usage{
tab1_fxn(tab_in, ds, grp, pp = 1, mp = 1, test = F, denom = F,
  header = "both")
}
\arguments{
\item{s}{survival object}

\item{CI}{include confidence interval (default TRUE)}

\item{plot.cens}{plot marks to indicate a censored subject (default TRUE)}

\item{surv.col}{colors to use for survival curves. Should match number of groups.}

\item{cens.col}{color of censor marks}

\item{cens.shape}{shape of censor marks}

\item{lty.est}{line type for survival curve}

\item{lty.ci}{line type for confidence interval}

\item{xlab}{label for x axis (default 'Time')}

\item{ylab}{label for y axis (default '')}

\item{main}{title for plot (default '')}

\item{cumProb}{plot cumulative probability instead of survival (default FALSE)}

\item{yTicks}{number of ticks on the y axis}

\item{addCounts}{add number still at risk above the x axis. This is an attempt to replicate SAS behavior. It does not always give good results (default FALSE)}

\item{bw}{print in black and white (default FALSE)}

\item{legend_title}{title to use for the legend}

\item{legend_pos}{where to position the legend}
}
\examples{

datas <- data.frame(grp=rep(c("A", "B", "C"), rep(100, 3)), 
var1 = rnorm(300, 0,1) + rep(1:3, rep(100, 3)), 
var2 = sample(c("Yes", "No"), 300, replace=T), 
var3 = sample(c("X", "Y", "Z"), 300, replace=T))
tab_desc <- bind_rows(data.frame(var="var1", varnm="Variable 1", type="c"),
                      data.frame(var="var2", varnm="Variable 2", type="b", target="Yes"),
                      data.frame(var="var3", varnm="Variable 3", type="m"))
tab1 <- tab_desc \%>\% group_by(ordr=1:n()) \%>\% do(tab1_fxn(., datas, grp="grp", test=T)) \%>\% ungroup
tab1 \%>\% select(-ordr) \%>\% TIRC(rnames="Characteristic", rgroup_col="group")
grp_tirc(tab1 \%>\% select(-ordr), grp="grp", p="p")
grp_tirc(tab1 \%>\% select(-ordr, -N), grp="grp", p="p")

}
\keyword{ggplot}
\keyword{survival}
