% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIRC.R
\name{TIRC}
\alias{TIRC}
\alias{TIRC.default}
\alias{TIRC.table}
\alias{print.TIRC}
\title{HTML Table With Some Extras

generic function to print html table.}
\usage{
TIRC(x, ...)

\method{TIRC}{default}(x, title = "", header, align = paste(rep("l",
  ncol(x)), collapse = ""), halign = paste(rep("l", ncol(x)), collapse = ""),
  cgroup, n.cgroup, cgroup.just, rgroup, n.rgroup,
  rgroupCSSstyle = "font-weight: bold", rgroupCSSseparator = "", tspanner,
  n.tspanner,
  tspannerCSSstyle = "font-weight: 900; text-transform:capitalize; text-align: center;",
  tspannerCSSseparator = "border-top: 1px solid grey;", rowlabel = title,
  rowlabel.pos = "bottom", headLines = "single",
  compatibility = "LibreOffice", rnames, caption, caption.loc = "top",
  tfoot, label, zebra = F, highrows, TURK = F, rgroup_col, ...)

\method{TIRC}{table}(x, title = "", header, align = paste(rep("l", ncol(x)),
  collapse = ""), halign = paste(rep("l", ncol(x)), collapse = ""), cgroup,
  n.cgroup, cgroup.just, rgroup, n.rgroup,
  rgroupCSSstyle = "font-weight: bold", rgroupCSSseparator = "", tspanner,
  n.tspanner,
  tspannerCSSstyle = "font-weight: 900; text-transform:capitalize; text-align: center;",
  tspannerCSSseparator = "border-top: 1px solid grey;", rowlabel = title,
  rowlabel.pos = "bottom", headLines = "single",
  compatibility = "LibreOffice", rnames, caption, caption.loc = "top",
  tfoot, label, zebra = F, highrows, TURK = F, ...)

\method{print}{TIRC}(x, useViewer, ...)
}
\arguments{
\item{x}{Matrix, data.frame, or table with the data}

\item{title}{String containing title for table}

\item{align}{defines the vertical alignment within cells (default All left alligned)}

\item{halign}{defines the horizontal alignment within cells (default All lower aligned)}

\item{cgroup}{defines the groupings for columns}

\item{n.cgroup}{defines the number of columns each cgroup should span. Must sum to total number of columns}

\item{cgroup.just}{defines horizontal alignment in cell for cgroup}

\item{rgroup}{defines the groupings for rows.}

\item{n.rgroup}{defines the number of rows each rgroup should span.}

\item{rgroupCSSstyle}{adds css to rgroup}

\item{rgroupCSSseparator}{css separator for rgroups}

\item{tspanner}{table spanner}

\item{n.tspanner}{counts for each table spanner}

\item{tspannerCSSseparator}{css to be added between table spanners}

\item{rowlabel}{string to be placed above rownames}

\item{rowlabel.pos}{defines the position of the row label}

\item{headLines}{defines how many horizontal lines start the table}

\item{compatibility}{some html tweaks for different systems}

\item{rnames}{defines row names. If a single string is given and it matches a column name then that column is used.}

\item{caption}{string used as caption}

\item{tfoot}{string used as footer for table}

\item{label}{adds a css id to table}

\item{zebra}{highlights alternate rows if TRUE (default FALSE)}

\item{highrows}{logical indicating rows to highlight or string indicating column containing highlight column}

\item{rgroup_col}{name of column to use instead of rgroup and n.rgroup}
}
\examples{

tab_data <- data.frame(rgrp=c("Early", "Early", "Late"),
                       rnms = c("A", "B", "C"), 
                       Red=c("10 (40\%)", "10 (40\%)", "5 (20\%)"), 
                       Green=c("20 (50\%)", "16 (40\%)", "4 (10\%)"))

TIRC(tab_data, rnames=F)
TIRC(tab_data, rnames="rnms")
TIRC(tab_data, rnames="rnms", rgroup_col = "rgrp")

tab_data2 <- data.frame(col=sample(c("Red", "Green"), 100, replace=T), 
                        shape=sample(c("Round", "Square"), 100, replace=T))

tab <- with(tab_data2, table(col, shape))
TIRC(tab)
}
\keyword{html}
\keyword{htmlTable}
\keyword{table}
