% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{tab1_fxn}
\alias{tab1_fxn}
\title{Table Maker


generic function to create a summary table of given variables.}
\usage{
tab1_fxn(tab_in, ds, grp, pp = 1, mp = 1, test = F, denom = F,
  header = "both")
}
\arguments{
\item{tab_in}{data.frame with variables var, varnm, and type. optional variables group, target, test_interval, fisher}

\item{ds}{dataset with variables referenced in tab_in}

\item{grp}{name of variable to be used for grouping}

\item{pp}{rounding digits for proportions}

\item{mp}{rounding digits for mean/sd}

\item{test}{boolean to request test of group differences}

\item{denom}{boolean to request display of denominator in proportions}

\item{header}{choose header for summary columns. "both" - "Percent (n) or Mean (SD)", "msd" - "Mean (SD)", np - "Percent (n)"}
}
\examples{

datas <- data.frame(grp=rep(c("A", "B", "C"), rep(100, 3)), 
var1 = rnorm(300, 0,1) + rep(1:3, rep(100, 3)), 
var2 = sample(c("Yes", "No"), 300, replace=T), 
var3 = sample(c("X", "Y", "Z"), 300, replace=T))
tab_desc <- bind_rows(data.frame(var="var1", varnm="Variable 1", type="c"),
                      data.frame(var="var2", varnm="Variable 2", type="b", target="Yes"),
                      data.frame(var="var3", varnm="Variable 3", type="m"))
tab1 <- tab_desc \%>\% group_by(ordr=1:n()) \%>\% do(tab1_fxn(., datas, grp="grp", test=T)) \%>\% ungroup
tab1 \%>\% select(-ordr) \%>\% TIRC(rnames="Characteristic", rgroup_col="group")
grp_tirc(tab1 \%>\% select(-ordr), grp="grp", p="p")
grp_tirc(tab1 \%>\% select(-ordr, -N), grp="grp", p="p")

}
\keyword{print}
\keyword{summary}
\keyword{table}
