% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtRound.R
\name{txtRound}
\alias{txtRound}
\title{A Rounding Function}
\usage{
txtRound(x, digits = 3, p = F, html = T)
}
\arguments{
\item{p}{adds in a 'p' before number (default FALSE)}

\item{html}{breaks and special characters given in html (default TRUE).}
}
\description{
This function rounds a number and returns a string.
}
\examples{
txtRound(c(0.002, 2.3049))
## [1] "0.002" "2.305"

txtRound(c(0.000002, 2.3049), p=T)
## [1] "p&nbsp;&lt;&nbsp;0.001" "p&nbsp;=&nbsp;2.305" 

txtRound(c(0.000002, 2.3049), p=T, html=F)
## [1] "p < 0.001" "p = 2.305"
}
\keyword{html}
\keyword{round}
\keyword{string}

